+++ src/components/icon/icon.component.js
@@ -6,6 +6,7 @@ import StyledIcon from "./icon.style";
 import Tooltip from "../tooltip";
 import { filterStyledSystemMarginProps } from "../../style/utils";
 import Logger from "../../utils/logger";
+import { ICON_TOOLTIP_POSITIONS } from "./icon-config";
 
 let deprecatedWarnTriggered = false;
 
@@ -117,8 +118,6 @@ const Icon = React.forwardRef(
   }
 );
 
-const placements = ["top", "bottom", "left", "right"];
-
 Icon.propTypes = {
   ...marginPropTypes,
   /**
@@ -168,7 +167,7 @@ Icon.propTypes = {
   /** The message string to be displayed in the tooltip */
   tooltipMessage: PropTypes.string,
   /** The position to display the tooltip */
-  tooltipPosition: PropTypes.oneOf(placements),
+  tooltipPosition: PropTypes.oneOf(["top", "bottom", "left", "right"]),
   /** Control whether the tooltip is visible */
   tooltipVisible: PropTypes.bool,
   /** Override background color of the Tooltip, provide any color from palette or any valid css color value. */
@@ -181,7 +180,7 @@ Icon.propTypes = {
     const isValid =
       prop &&
       Array.isArray(prop) &&
-      prop.every((placement) => placements.includes(placement));
+      prop.every((placement) => ICON_TOOLTIP_POSITIONS.includes(placement));
 
     if (!prop || isValid) {
       return null;