+++ src/components/sidebar/__internal__/sidebar-header/sidebar-header.style.js
@@ -4,7 +4,7 @@ import baseTheme from "../../../../style/themes/base";
 const SidebarHeaderStyle = styled.div`
   background-color: ${({ theme }) => theme.colors.white};
   box-shadow: inset 0 -1px 0 0 ${({ theme }) => theme.disabled.border};
-  box-sizing: content-box;
+  box-sizing: border-box;
   width: 100%;
   color: ${({ theme }) => theme.text.color};
   transition: all 0.2s ease;