+++ src/components/sidebar/sidebar.stories.js
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
-import { boolean, select } from "@storybook/addon-knobs";
+import { boolean, select, text } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
-import OptionsHelper from "../../utils/helpers/options-helper";
 import Sidebar from ".";
 import Button from "../button";
+import { SIDEBAR_ALIGNMENTS, SIDEBAR_SIZES } from "./sidebar.config";
 
 export default {
   title: "Sidebar/Test",
@@ -20,20 +20,17 @@ export default {
 
 export const Default = () => {
   const [isOpen, setIsOpen] = useState(true);
+
   const enableBackgroundUI = boolean(
     "enableBackgroundUI",
     Sidebar.defaultProps.enableBackgroundUI
   );
   const position = select(
     "position",
-    OptionsHelper.alignBinary,
+    SIDEBAR_ALIGNMENTS,
     Sidebar.defaultProps.position
   );
-  const size = select(
-    "size",
-    OptionsHelper.sizesFull,
-    Sidebar.defaultProps.size
-  );
+  const size = select("size", SIDEBAR_SIZES, Sidebar.defaultProps.size);
 
   const onCancel = () => {
     setIsOpen(false);
@@ -43,6 +40,8 @@ export const Default = () => {
   const openSidebar = () => {
     setIsOpen(true);
   };
+
+  const headerChildren = text("header", "");
   return (
     <>
       <Button onClick={openSidebar}>Open sidebar</Button>
@@ -52,6 +51,7 @@ export const Default = () => {
         position={position}
         size={size}
         onCancel={onCancel}
+        header={headerChildren}
       >
         <div>
           <Button as="primary">Test</Button>