+++ src/components/switch/__internal__/switch-slider-panel.spec.js
@@ -1,6 +1,5 @@
 import React from "react";
 import TestRenderer from "react-test-renderer";
-import "jest-styled-components";
 import { css } from "styled-components";
 
 import {