+++ src/components/table/table-story-helpers/table-story-wrapper.component.js
@@ -10,14 +10,13 @@ import TextArea from "../../textarea";
 import DateInput from "../../date";
 import TextBox from "../../textbox";
 import { getCommonTextboxProps } from "../../textbox/textbox.stories";
-import OptionsHelper from "../../../utils/helpers/options-helper/options-helper";
+import { TABLE_INPUT_TYPES } from "../table.config";
 
 const pickInput = (name) => {
-  const { inputTypes } = OptionsHelper;
   switch (name) {
-    case inputTypes[1]:
+    case TABLE_INPUT_TYPES[1]:
       return <TextArea {...getCommonTextboxProps} />;
-    case inputTypes[2]:
+    case TABLE_INPUT_TYPES[2]:
       return <DateInput {...getCommonTextboxProps} />;
     default:
       return <TextBox {...getCommonTextboxProps} />;